requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
Markdown==3.5.1
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.2.3
python-dotenv